Bluetooth: Move HoG implementation to subsys/bluetooth/lib
This moves HID over GATT Service implementation under
subsys/bluetooth/lib and integrate with Kbuild to avoid build
problems.

Fixes #1166
