+++ samples/bluetooth/peripheral_hids/prj.conf
@@ -6,3 +6,4 @@ CONFIG_BT_DEVICE_NAME="Test HoG mouse"
 CONFIG_BT_DEVICE_APPEARANCE=962
 CONFIG_BT_BAS=y
 CONFIG_BT_DIS=y
+CONFIG_BT_HOG=y
