+++ samples/bluetooth/peripheral_hids/src/main.c
@@ -21,8 +21,7 @@
 #include <bluetooth/gatt.h>
 #include <bluetooth/bas.h>
 #include <bluetooth/dis.h>
-
-#include <gatt/hog.h>
+#include <bluetooth/hog.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN		(sizeof(DEVICE_NAME) - 1)
