+++ subsys/bluetooth/lib/Makefile
@@ -1,3 +1,4 @@
 obj-$(CONFIG_BT_BAS) += bas.o
 obj-$(CONFIG_BT_CTS) += cts.o
 obj-$(CONFIG_BT_DIS) += dis.o
+obj-$(CONFIG_BT_HOG) += hog.o
