+++ subsys/bluetooth/lib/hog.c
